+++ include/fsom/EffectClasses/Flanger.hpp
@@ -29,8 +29,8 @@ namespace fsom{
 
 class Flanger :public DSPEffect
 {
-	DelayBase<double> m_delayUnitL,m_delayUnitR;
-	Table<double> m_table;
+	DelayBase<float> m_delayUnitL,m_delayUnitR;
+	Table<float> m_table;
 	Phasor m_phasor;
 public:
 	Flanger(dspCreationStruct data );
