+++ include/fsom/EffectClasses/TransposeUnit.hpp
@@ -29,8 +29,8 @@ namespace fsom{
 class TransposeUnit 
 {
   const int m_delaySize;
-  DelayBase<double> m_delayUnitL,m_delayUnitR;
-  Table<double> m_table;
+  DelayBase<float> m_delayUnitL,m_delayUnitR;
+  Table<float> m_table;
   Phasor m_phasor;
   int m_proportion;
