+++ src/GrainStream.cpp
@@ -69,20 +69,20 @@ GrainStream::GrainStream():
   m_grainSize(44100),
   m_nextSpawn(44100),
   m_grainRate(1),
-  m_window(TablePtr(new Table<double>(512))),
+  m_window(TablePtr(new Table<float>(512))),
   m_filepath("")
 {
   m_window->fill_hann();
 //   spawn();
   
-  TablePtr t1 = TablePtr(new Table<double>(44100));
-  TablePtr t2 = TablePtr(new Table<double>(44100));
-  MultiTablePtr mt(new MultiChannelBuffer(TablePtr(new Table<double>(44100)));
+  TablePtr t1 = TablePtr(new Table<float>(44100));
+  TablePtr t2 = TablePtr(new Table<float>(44100));
+  MultiTablePtr mt(new MultiChannelBuffer(TablePtr(new Table<float>(44100)));
   mt->push_back(t1);
   mt->push_back(t2);
   MultiTablePtr mt;
-  TablePtr t1 = TablePtr(new Table<double>(44100*5));
-  TablePtr t2 = TablePtr(new Table<double>(44100*5));
+  TablePtr t1 = TablePtr(new Table<float>(44100*5));
+  TablePtr t2 = TablePtr(new Table<float>(44100*5));
   m_table = MultiTablePtr(new MultiTableBuffer());
   m_table->push_back(t1);
   m_table->push_back(t2);
